+++ src/makefile
@@ -307,8 +307,8 @@ UNIX_OBJS=\
 	timeout.$(O) \
 	io.$(O) \
 	usocket.$(O) \
-	unixtcp.$(O) \
-	unixudp.$(O) \
+	unixstream.$(O) \
+	unixdgram.$(O) \
 	compat.$(O) \
 	unix.$(O)
