+++ src/unixstream.c
@@ -1,9 +1,8 @@
 /*=========================================================================*\
-* Unix domain socket udp submodule
+* Unix domain socket stream sub module
 * LuaSocket toolkit
 \*=========================================================================*/
 #include <string.h>
-#include <stdlib.h>
 
 #include "lua.h"
 #include "lauxlib.h"
@@ -12,81 +11,86 @@
 #include "auxiliar.h"
 #include "socket.h"
 #include "options.h"
-#include "unix.h"
+#include "unixstream.h"
 #include <sys/un.h>
-
-#define UNIXUDP_DATAGRAMSIZE 8192
 
 /*=========================================================================*\
 * Internal function prototypes
 \*=========================================================================*/
 static int global_create(lua_State *L);
 static int meth_connect(lua_State *L);
+static int meth_listen(lua_State *L);
 static int meth_bind(lua_State *L);
 static int meth_send(lua_State *L);
+static int meth_shutdown(lua_State *L);
 static int meth_receive(lua_State *L);
+static int meth_accept(lua_State *L);
 static int meth_close(lua_State *L);
 static int meth_setoption(lua_State *L);
 static int meth_settimeout(lua_State *L);
-static int meth_gettimeout(lua_State *L);
 static int meth_getfd(lua_State *L);
 static int meth_setfd(lua_State *L);
 static int meth_dirty(lua_State *L);
-static int meth_receivefrom(lua_State *L);
-static int meth_sendto(lua_State *L);
+static int meth_getstats(lua_State *L);
+static int meth_setstats(lua_State *L);
 static int meth_getsockname(lua_State *L);
 
-static const char *unixudp_tryconnect(p_unix un, const char *path);
-static const char *unixudp_trybind(p_unix un, const char *path);
+static const char *unixstream_tryconnect(p_unix un, const char *path);
+static const char *unixstream_trybind(p_unix un, const char *path);
 
-/* unixudp object methods */
-static luaL_Reg unixudp_methods[] = {
+/* unixstream object methods */
+static luaL_Reg unixstream_methods[] = {
     {"__gc",        meth_close},
     {"__tostring",  auxiliar_tostring},
+    {"accept",      meth_accept},
     {"bind",        meth_bind},
     {"close",       meth_close},
     {"connect",     meth_connect},
     {"dirty",       meth_dirty},
     {"getfd",       meth_getfd},
-    {"send",        meth_send},
-    {"sendto",      meth_sendto},
+    {"getstats",    meth_getstats},
+    {"setstats",    meth_setstats},
+    {"listen",      meth_listen},
     {"receive",     meth_receive},
-    {"receivefrom", meth_receivefrom},
+    {"send",        meth_send},
     {"setfd",       meth_setfd},
     {"setoption",   meth_setoption},
     {"setpeername", meth_connect},
     {"setsockname", meth_bind},
     {"getsockname", meth_getsockname},
     {"settimeout",  meth_settimeout},
-    {"gettimeout",  meth_gettimeout},
+    {"shutdown",    meth_shutdown},
     {NULL,          NULL}
 };
 
 /* socket option handlers */
 static t_opt optset[] = {
+    {"keepalive",   opt_set_keepalive},
     {"reuseaddr",   opt_set_reuseaddr},
+    {"linger",      opt_set_linger},
     {NULL,          NULL}
 };
 
 /* functions in library namespace */
 static luaL_Reg func[] = {
-    {"udp", global_create},
+    {"stream", global_create},
     {NULL, NULL}
 };
 
 /*-------------------------------------------------------------------------*\
 * Initializes module
 \*-------------------------------------------------------------------------*/
-int unixudp_open(lua_State *L)
+int unixstream_open(lua_State *L)
 {
     /* create classes */
-    auxiliar_newclass(L, "unixudp{connected}", unixudp_methods);
-    auxiliar_newclass(L, "unixudp{unconnected}", unixudp_methods);
+    auxiliar_newclass(L, "unixstream{master}", unixstream_methods);
+    auxiliar_newclass(L, "unixstream{client}", unixstream_methods);
+    auxiliar_newclass(L, "unixstream{server}", unixstream_methods);
+
     /* create class groups */
-    auxiliar_add2group(L, "unixudp{connected}",   "unixudp{any}");
-    auxiliar_add2group(L, "unixudp{unconnected}", "unixudp{any}");
-    auxiliar_add2group(L, "unixudp{connected}",   "select{able}");
-    auxiliar_add2group(L, "unixudp{unconnected}", "select{able}");
+    auxiliar_add2group(L, "unixstream{master}", "unixstream{any}");
+    auxiliar_add2group(L, "unixstream{client}", "unixstream{any}");
+    auxiliar_add2group(L, "unixstream{server}", "unixstream{any}");
 
     luaL_setfuncs(L, func, 0);
     return 0;
@@ -95,139 +99,34 @@ int unixudp_open(lua_State *L)
 /*=========================================================================*\
 * Lua methods
 \*=========================================================================*/
-static const char *unixudp_strerror(int err)
-{
-    /* a 'closed' error on an unconnected means the target address was not
-     * accepted by the transport layer */
-    if (err == IO_CLOSED) return "refused";
-    else return socket_strerror(err);
-}
-
-static int meth_send(lua_State *L)
-{
-    p_unix un = (p_unix) auxiliar_checkclass(L, "unixudp{connected}", 1);
-    p_timeout tm = &un->tm;
-    size_t count, sent = 0;
-    int err;
-    const char *data = luaL_checklstring(L, 2, &count);
-    timeout_markstart(tm);
-    err = socket_send(&un->sock, data, count, &sent, tm);
-    if (err != IO_DONE) {
-        lua_pushnil(L);
-        lua_pushstring(L, unixudp_strerror(err));
-        return 2;
-    }
-    lua_pushnumber(L, (lua_Number) sent);
-    return 1;
-}
-
 /*-------------------------------------------------------------------------*\
-* Send data through unconnected unixudp socket
+* Just call buffered IO methods
 \*-------------------------------------------------------------------------*/
-static int meth_sendto(lua_State *L)
-{
-    p_unix un = (p_unix) auxiliar_checkclass(L, "unixudp{unconnected}", 1);
-    size_t count, sent = 0;
-    const char *data = luaL_checklstring(L, 2, &count);
-    const char *path = luaL_checkstring(L, 3);
-    p_timeout tm = &un->tm;
-    int err;
-    struct sockaddr_un remote;
-    size_t len = strlen(path);
-
-    if (len >= sizeof(remote.sun_path)) {
-		lua_pushnil(L);
-		lua_pushstring(L, "path too long");
-		return 2;
-	}
-
-    memset(&remote, 0, sizeof(remote));
-    strcpy(remote.sun_path, path);
-    remote.sun_family = AF_UNIX;
-    timeout_markstart(tm);
-#ifdef UNIX_HAS_SUN_LEN
-    remote.sun_len = sizeof(remote.sun_family) + sizeof(remote.sun_len)
-        + len + 1;
-    err = socket_sendto(&un->sock, data, count, &sent, (SA *) &remote, remote.sun_len, tm);
-#else
-    err = socket_sendto(&un->sock, data, count, &sent, (SA *) &remote,
-		   	sizeof(remote.sun_family) + len, tm);
-#endif
-    if (err != IO_DONE) {
-        lua_pushnil(L);
-        lua_pushstring(L, unixudp_strerror(err));
-        return 2;
-    }
-    lua_pushnumber(L, (lua_Number) sent);
-    return 1;
+static int meth_send(lua_State *L) {
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{client}", 1);
+    return buffer_meth_send(L, &un->buf);
 }
 
 static int meth_receive(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
-    char buf[UNIXUDP_DATAGRAMSIZE];
-    size_t got, wanted = (size_t) luaL_optnumber(L, 2, sizeof(buf));
-    char *dgram = wanted > sizeof(buf)? (char *) malloc(wanted): buf;
-    int err;
-    p_timeout tm = &un->tm;
-    timeout_markstart(tm);
-    if (!dgram) {
-        lua_pushnil(L);
-        lua_pushliteral(L, "out of memory");
-        return 2;
-    }
-    err = socket_recv(&un->sock, dgram, wanted, &got, tm);
-    /* Unlike TCP, recv() of zero is not closed, but a zero-length packet. */
-    if (err != IO_DONE && err != IO_CLOSED) {
-        lua_pushnil(L);
-        lua_pushstring(L, unixudp_strerror(err));
-        if (wanted > sizeof(buf)) free(dgram);
-        return 2;
-    }
-    lua_pushlstring(L, dgram, got);
-    if (wanted > sizeof(buf)) free(dgram);
-    return 1;
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{client}", 1);
+    return buffer_meth_receive(L, &un->buf);
 }
 
-/*-------------------------------------------------------------------------*\
-* Receives data and sender from a UDP socket
-\*-------------------------------------------------------------------------*/
-static int meth_receivefrom(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkclass(L, "unixudp{unconnected}", 1);
-    char buf[UNIXUDP_DATAGRAMSIZE];
-    size_t got, wanted = (size_t) luaL_optnumber(L, 2, sizeof(buf));
-    char *dgram = wanted > sizeof(buf)? (char *) malloc(wanted): buf;
-    struct sockaddr_un addr;
-    socklen_t addr_len = sizeof(addr);
-    int err;
-    p_timeout tm = &un->tm;
-    timeout_markstart(tm);
-    if (!dgram) {
-        lua_pushnil(L);
-        lua_pushliteral(L, "out of memory");
-        return 2;
-    }
-    err = socket_recvfrom(&un->sock, dgram, wanted, &got, (SA *) &addr,
-            &addr_len, tm);
-    /* Unlike TCP, recv() of zero is not closed, but a zero-length packet. */
-    if (err != IO_DONE && err != IO_CLOSED) {
-        lua_pushnil(L);
-        lua_pushstring(L, unixudp_strerror(err));
-        if (wanted > sizeof(buf)) free(dgram);
-        return 2;
-    }
+static int meth_getstats(lua_State *L) {
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{client}", 1);
+    return buffer_meth_getstats(L, &un->buf);
+}
 
-    lua_pushlstring(L, dgram, got);
-	/* the path may be empty, when client send without bind */
-    lua_pushstring(L, addr.sun_path);
-    if (wanted > sizeof(buf)) free(dgram);
-    return 2;
+static int meth_setstats(lua_State *L) {
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{client}", 1);
+    return buffer_meth_setstats(L, &un->buf);
 }
 
 /*-------------------------------------------------------------------------*\
 * Just call option handler
 \*-------------------------------------------------------------------------*/
 static int meth_setoption(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     return opt_meth_setoption(L, optset, &un->sock);
 }
 
@@ -235,29 +134,56 @@ static int meth_setoption(lua_State *L) {
 * Select support methods
 \*-------------------------------------------------------------------------*/
 static int meth_getfd(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     lua_pushnumber(L, (int) un->sock);
     return 1;
 }
 
 /* this is very dangerous, but can be handy for those that are brave enough */
 static int meth_setfd(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     un->sock = (t_socket) luaL_checknumber(L, 2);
     return 0;
 }
 
 static int meth_dirty(lua_State *L) {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
-    (void) un;
-    lua_pushboolean(L, 0);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
+    lua_pushboolean(L, !buffer_isempty(&un->buf));
     return 1;
+}
+
+/*-------------------------------------------------------------------------*\
+* Waits for and returns a client object attempting connection to the
+* server object
+\*-------------------------------------------------------------------------*/
+static int meth_accept(lua_State *L) {
+    p_unix server = (p_unix) auxiliar_checkclass(L, "unixstream{server}", 1);
+    p_timeout tm = timeout_markstart(&server->tm);
+    t_socket sock;
+    int err = socket_accept(&server->sock, &sock, NULL, NULL, tm);
+    /* if successful, push client socket */
+    if (err == IO_DONE) {
+        p_unix clnt = (p_unix) lua_newuserdata(L, sizeof(t_unix));
+        auxiliar_setclass(L, "unixstream{client}", -1);
+        /* initialize structure fields */
+        socket_setnonblocking(&sock);
+        clnt->sock = sock;
+        io_init(&clnt->io, (p_send)socket_send, (p_recv)socket_recv,
+                (p_error) socket_ioerror, &clnt->sock);
+        timeout_init(&clnt->tm, -1, -1);
+        buffer_init(&clnt->buf, &clnt->io, &clnt->tm);
+        return 1;
+    } else {
+        lua_pushnil(L);
+        lua_pushstring(L, socket_strerror(err));
+        return 2;
+    }
 }
 
 /*-------------------------------------------------------------------------*\
 * Binds an object to an address
 \*-------------------------------------------------------------------------*/
-static const char *unixudp_trybind(p_unix un, const char *path) {
+static const char *unixstream_trybind(p_unix un, const char *path) {
     struct sockaddr_un local;
     size_t len = strlen(path);
     int err;
@@ -278,11 +204,10 @@ static const char *unixudp_trybind(p_unix un, const char *path) {
     return socket_strerror(err);
 }
 
-static int meth_bind(lua_State *L)
-{
-    p_unix un = (p_unix) auxiliar_checkclass(L, "unixudp{unconnected}", 1);
+static int meth_bind(lua_State *L) {
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{master}", 1);
     const char *path =  luaL_checkstring(L, 2);
-    const char *err = unixudp_trybind(un, path);
+    const char *err = unixstream_trybind(un, path);
     if (err) {
         lua_pushnil(L);
         lua_pushstring(L, err);
@@ -294,7 +219,7 @@ static int meth_bind(lua_State *L)
 
 static int meth_getsockname(lua_State *L)
 {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     struct sockaddr_un peer = {0};
     socklen_t peer_len = sizeof(peer);
 
@@ -309,9 +234,9 @@ static int meth_getsockname(lua_State *L)
 }
 
 /*-------------------------------------------------------------------------*\
-* Turns a master unixudp object into a client object.
+* Turns a master unixstream object into a client object.
 \*-------------------------------------------------------------------------*/
-static const char *unixudp_tryconnect(p_unix un, const char *path)
+static const char *unixstream_tryconnect(p_unix un, const char *path)
 {
     struct sockaddr_un remote;
     int err;
@@ -335,16 +260,16 @@ static const char *unixudp_tryconnect(p_unix un, const char *path)
 
 static int meth_connect(lua_State *L)
 {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{master}", 1);
     const char *path =  luaL_checkstring(L, 2);
-    const char *err = unixudp_tryconnect(un, path);
+    const char *err = unixstream_tryconnect(un, path);
     if (err) {
         lua_pushnil(L);
         lua_pushstring(L, err);
         return 2;
     }
-    /* turn unconnected object into a connected object */
-    auxiliar_setclass(L, "unixudp{connected}", 1);
+    /* turn master object into a client object */
+    auxiliar_setclass(L, "unixstream{client}", 1);
     lua_pushnumber(L, 1);
     return 1;
 }
@@ -354,8 +279,41 @@ static int meth_connect(lua_State *L)
 \*-------------------------------------------------------------------------*/
 static int meth_close(lua_State *L)
 {
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     socket_destroy(&un->sock);
+    lua_pushnumber(L, 1);
+    return 1;
+}
+
+/*-------------------------------------------------------------------------*\
+* Puts the sockt in listen mode
+\*-------------------------------------------------------------------------*/
+static int meth_listen(lua_State *L)
+{
+    p_unix un = (p_unix) auxiliar_checkclass(L, "unixstream{master}", 1);
+    int backlog = (int) luaL_optnumber(L, 2, 32);
+    int err = socket_listen(&un->sock, backlog);
+    if (err != IO_DONE) {
+        lua_pushnil(L);
+        lua_pushstring(L, socket_strerror(err));
+        return 2;
+    }
+    /* turn master object into a server object */
+    auxiliar_setclass(L, "unixstream{server}", 1);
+    lua_pushnumber(L, 1);
+    return 1;
+}
+
+/*-------------------------------------------------------------------------*\
+* Shuts the connection down partially
+\*-------------------------------------------------------------------------*/
+static int meth_shutdown(lua_State *L)
+{
+    /* SHUT_RD,  SHUT_WR,  SHUT_RDWR  have  the value 0, 1, 2, so we can use method index directly */
+    static const char* methods[] = { "receive", "send", "both", NULL };
+    p_unix stream = (p_unix) auxiliar_checkclass(L, "unixstream{client}", 1);
+    int how = luaL_checkoption(L, 2, "both", methods);
+    socket_shutdown(&stream->sock, how);
     lua_pushnumber(L, 1);
     return 1;
 }
@@ -363,34 +321,26 @@ static int meth_close(lua_State *L)
 /*-------------------------------------------------------------------------*\
 * Just call tm methods
 \*-------------------------------------------------------------------------*/
-static int meth_settimeout(lua_State *L)
-{
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
+static int meth_settimeout(lua_State *L) {
+    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixstream{any}", 1);
     return timeout_meth_settimeout(L, &un->tm);
-}
-
-static int meth_gettimeout(lua_State *L)
-{
-    p_unix un = (p_unix) auxiliar_checkgroup(L, "unixudp{any}", 1);
-    return timeout_meth_gettimeout(L, &un->tm);
 }
 
 /*=========================================================================*\
 * Library functions
 \*=========================================================================*/
 /*-------------------------------------------------------------------------*\
-* Creates a master unixudp object
+* Creates a master unixstream object
 \*-------------------------------------------------------------------------*/
-static int global_create(lua_State *L)
-{
+static int global_create(lua_State *L) {
     t_socket sock;
-    int err = socket_create(&sock, AF_UNIX, SOCK_DGRAM, 0);
+    int err = socket_create(&sock, AF_UNIX, SOCK_STREAM, 0);
     /* try to allocate a system socket */
     if (err == IO_DONE) {
-        /* allocate unixudp object */
+        /* allocate unixstream object */
         p_unix un = (p_unix) lua_newuserdata(L, sizeof(t_unix));
         /* set its type as master object */
-        auxiliar_setclass(L, "unixudp{unconnected}", -1);
+        auxiliar_setclass(L, "unixstream{master}", -1);
         /* initialize remaining structure fields */
         socket_setnonblocking(&sock);
         un->sock = sock;
