+++ test/unixdgramclnt.lua
@@ -1,7 +1,7 @@
 socket = require"socket"
 socket.unix = require"socket.unix"
-c = assert(socket.unix.udp())
-c:bind("/tmp/bar")
+c = assert(socket.unix.dgram())
+print(c:bind("/tmp/bar"))
 while 1 do
     local l = io.read("*l")
     assert(c:sendto(l, "/tmp/foo"))
