+++ test/unixdgramsrvr.lua
@@ -1,6 +1,6 @@
     socket = require"socket"
     socket.unix = require"socket.unix"
-    u = assert(socket.unix.udp())
+    u = assert(socket.unix.dgram())
     assert(u:bind("/tmp/foo"))
     while 1 do
 		x, r = assert(u:receivefrom())
