+++ test/unixstreamclnt.lua
@@ -1,6 +1,6 @@
 socket = require"socket"
 socket.unix = require"socket.unix"
-c = assert(socket.unix.tcp())
+c = assert(socket.unix.stream())
 assert(c:connect("/tmp/foo"))
 while 1 do
     local l = io.read()
