+++ test/unixstreamsrvr.lua
@@ -1,6 +1,6 @@
     socket = require"socket"
     socket.unix = require"socket.unix"
-    u = assert(socket.unix.tcp())
+    u = assert(socket.unix.stream())
     assert(u:bind("/tmp/foo"))
     assert(u:listen())
     c = assert(u:accept())
